fix(test): split test_error_tracking.py to meet LOC ceiling
CI failed due to test_error_tracking.py exceeding 650 line limit (739 lines).
Extracted 5 get_all_errors_status() edge case tests to new file
test_error_tracking_edge_cases.py.

Changes:
- test_error_tracking.py: 739 → 611 lines (within limit)
- test_error_tracking_edge_cases.py: new file with 5 edge case tests

All 42 tests pass. No functional changes.
